requirements.txt
click==7.1.2
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
numpy==1.16.6
Pillow==6.2.2
opencv-python==4.2.0.32
python-dateutil==2.5.0

